// src/state/routing/types.ts
export enum TradeState {
  LOADING,
  INVALID,
  STALE,
  NO_ROUTE_FOUND,
  VALID,
}
// ...
export enum TradeType {
  EXACT_INPUT = 'EXACT_INPUT',
  EXACT_OUTPUT = 'EXACT_OUTPUT',
}

export enum PlatformType {
  AGGREGATOR = 'Soroswap Aggregator',
  ROUTER = 'Soroswap AMM',
  STELLAR_CLASSIC = 'SDEX',
}

export type InterfaceTrade = {
  inputAmount: CurrencyAmount | undefined;
  outputAmount: CurrencyAmount | undefined;
  tradeType: TradeType | undefined;
  path: string[] | undefined;
  distribution?: DexDistribution[] | undefined;
  priceImpact?: Percent | Number | undefined;
  [x: string]: any;
  platform?: PlatformType;
};